TaskThree/mavenCalculatorProject/src/main/java/by/katomakhina/parse/Parser.java
package by.katomakhina.parse;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

public final class Parser {
    public static String[] parse(String expression) {
        Pattern pattern = Pattern.compile("^[0-9]{1,10}\\.?([0-9]?){1,10}[+\\-/*][0-9]{1,10}\\.?([0-9]?){1,10}$");
        Matcher matcher = pattern.matcher(expression);
        if (matcher.find()) {
            String[] a = matcher.group().split("[+\\-/*]");
            return a;
        } else {
            return null;
        }
    }

    public static String getSignFromString(String expression){
        if (expression.contains("+")) {
            return "+";
        }
        if (expression.contains("-")) {
            return "-";
        }
        if (expression.contains("*")) {
            return "*";
        }
        if (expression.contains("/")) {
            return "/";
        }
        return null;
    }


    public Parser(){}
}
